chore: improve scripts to clean repo
